refactor: rename components and hooks to kebab-case convention
- Rename DevtoolsApp.tsx → dev-tools-app.tsx
- Rename OptionsApp.tsx → options-app.tsx
- Rename SidePanelApp.tsx → side-panel-app.tsx
- Rename BaseLayout.tsx → base-layout.tsx
- Rename useAsyncTask.ts → use-async-task.ts
- Rename useStorage.ts → use-storage.ts

This establishes consistent kebab-case naming throughout the codebase
for better maintainability and follows common React project conventions.
